# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic Control Plane API
  version: 0.2.0
  summary: Policy, approval, execution, and audit APIs for governed agent actions
  description: |
    v0.2 extends the base control-plane lifecycle with richer, domain-specific schemas for:
      - Git and pull request actions
      - Deploy and cloud actions
      - Email and calendar actions
      - Payment and procurement actions

    Core lifecycle:
      intent -> policy decision -> approval (optional) -> execution grant -> execution -> trace
servers:
  - url: https://api.withaileron.ai
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /v1/users/me/passphrase:
    post:
      tags: [Users]
      summary: Set or rotate vault passphrase
      description: |
        Stores the client-provided salt and KEK verification blob. The client
        derives the KEK locally using Argon2id, encrypts a known verification
        constant, and sends only the salt and encrypted blob — the passphrase
        and KEK never leave the browser. On subsequent calls, the existing key
        material is rotated (callers must re-encrypt vault secrets separately).
      operationId: setPassphrase
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPassphraseRequest"
      responses:
        "200":
          description: Passphrase set successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassphraseSaltResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me/passphrase/verification:
    get:
      tags: [Users]
      summary: Get KEK verification blob
      description: |
        Returns the encrypted verification blob for client-side passphrase
        verification. The client derives the KEK locally from the passphrase
        and salt, then decrypts this blob to check correctness. The server
        never sees the passphrase or KEK.
      operationId: getPassphraseVerification
      responses:
        "200":
          description: Verification blob
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassphraseVerificationResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me/passphrase/salt:
    get:
      tags: [Users]
      summary: Get passphrase salt
      description: |
        Returns the Argon2id salt for the current user's passphrase, allowing
        client-side KEK derivation. Also indicates whether the user has set a
        passphrase.
      operationId: getPassphraseSalt
      responses:
        "200":
          description: Salt and passphrase status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PassphraseSaltResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me/passphrase/unlock:
    post:
      tags: [Users]
      summary: Unlock vault with passphrase-derived KEK
      description: |
        The client derives the KEK locally from the passphrase and stored salt
        using Argon2id, then sends the 32-byte KEK to the server over HTTPS.
        The server verifies it against the stored verification blob, and on
        success caches the KEK in memory for the session TTL. Tokens can only
        be encrypted or decrypted while the vault is unlocked.
      operationId: unlockVault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnlockVaultRequest"
      responses:
        "200":
          description: Vault unlocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStatusResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /v1/users/me/passphrase/lock:
    post:
      tags: [Users]
      summary: Lock vault and clear cached KEK
      description: |
        Clears the user's cached KEK from server memory. Connected account
        operations will return 423 until the vault is unlocked again.
      operationId: lockVault
      responses:
        "200":
          description: Vault locked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/users/me/vault/status:
    get:
      tags: [Users]
      summary: Get vault lock status
      description: |
        Returns whether the vault is locked or unlocked, the session expiry
        time, and whether the user has set a passphrase.
      operationId: getVaultStatus
      responses:
        "200":
          description: Vault status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultStatusResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true
            request_id: { type: string }
...
    SetPassphraseRequest:
      type: object
      required: [salt, kek_verification]
      properties:
        salt:
          type: string
          format: byte
          description: Base64-encoded 16-byte Argon2id salt generated by the client
        kek_verification:
          type: string
          format: byte
          description: Verification constant encrypted with the client-derived KEK (base64)
    PassphraseVerificationResponse:
      type: object
      required: [has_passphrase]
      properties:
        has_passphrase:
          type: boolean
          description: Whether the user has set a vault passphrase
        kek_verification:
          type: string
          format: byte
          description: Encrypted verification blob for client-side decryption (omitted when no passphrase is set)
    PassphraseSaltResponse:
      type: object
      required: [has_passphrase]
      properties:
        has_passphrase:
          type: boolean
          description: Whether the user has set a vault passphrase
        salt:
          type: string
          format: byte
          description: Base64-encoded Argon2id salt (omitted when no passphrase is set)
    UnlockVaultRequest:
      type: object
      required: [kek]
      properties:
        kek:
          type: string
          format: byte
          description: Base64-encoded 32-byte KEK derived from passphrase via Argon2id
    VaultStatusResponse:
      type: object
      required: [locked, has_passphrase]
      properties:
        locked:
          type: boolean
          description: Whether the vault is currently locked (no active KEK session)
        has_passphrase:
          type: boolean
          description: Whether the user has set a vault passphrase
        expires_at:
          type: string
          format: date-time
          description: When the current KEK session expires (omitted when locked)
...